soma de numero com virgula
import java.util.Scanner; // Importa a classe Scanner para ler dados do usuário

public class Main {
    public static void main(String[] args) {

        Scanner scanner = new Scanner(System.in);
        // Cria um objeto Scanner chamado "scanner" para ler entradas do teclado

        System.out.print("Digite o primeiro numero: ");
        // Mostra a mensagem pedindo para o usuário digitar o primeiro número
        float numero1 = scanner.nextFloat();
        // Lê o número decimal digitado pelo usuário e armazena na variável "numero1"

        System.out.print("Digite o segundo numero: ");
        // Mostra a mensagem pedindo para o usuário digitar o segundo número
        float numero2 = scanner.nextFloat();
        // Lê o segundo número decimal digitado e armazena na variável "numero2"

        float resultado = numero1 + numero2;
        // Soma os dois números e armazena o resultado na variável "resultado"

        System.out.println("Resultado: " + resultado);
        // Mostra na tela o resultado da soma
    }
}

+++ src/Main.java
@@ -5,12 +5,12 @@ public class Main {
         Scanner scanner = new Scanner(System.in);
 
         System.out.print("Digite o primeiro numero: ");
-        int numero1 = scanner.nextInt();
+        float numero1 = scanner.nextFloat();
 
         System.out.print("Digite o segundo numero: ");
-        int numero2 = scanner.nextInt();
+        float numero2 = scanner.nextFloat();
 
-        int resultado = numero1 + numero2;
+        float resultado = numero1 + numero2;
         System.out.println("Resultado: " + resultado);
     }
 }
